working backend checkpoint
