+++ src/templates/.flaskenv
@@ -1,0 +1,2 @@
+FLASK_APP=app.py
+FLASK_DEBUG=true
