+++ src/templates/shower_thoughts.csv
@@ -1,0 +1,1 @@
+1,N/A,N/A,1,N/A,1,N/A,N/A,N/A,new one,N/A
