+++ src/templates/static/package.json
@@ -1,6 +1,7 @@
 {
   "name": "showerthoughts-static",
   "version": "1.0.0",
+  "proxy": "http://127.0.0.1:5000",
   "description": "",
   "main": "index.jsx",
   "scripts": {
@@ -20,6 +21,7 @@
     "webpack-cli": "^5.0.1"
   },
   "dependencies": {
+    "axios": "^1.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-force-graph-3d": "^1.21.17",
